% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_races.R
\name{get_races}
\alias{get_races}
\title{Get all races}
\usage{
get_races()
}
\value{
Returns a tibble of race results
}
\description{
Returns a tibble of all race results from \url{road-results.com}.
}
\examples{
get_races()
}
